+++ .travis.yml
@@ -4,3 +4,5 @@ node_js:
   - '6'
   - '8'
   - '10'
+after_success:
+  - npm run coveralls
