+++ package.json
@@ -13,7 +13,9 @@
     "node": ">=4.8.3"
   },
   "scripts": {
-    "test": "mocha"
+    "test": "mocha",
+    "cover": "nyc mocha",
+    "coveralls": "nyc --reporter=lcov mocha && cat ./coverage/lcov.info | coveralls"
   },
   "files": [
     "index.js"
@@ -33,9 +35,11 @@
     "vinyl": "^2.2.0"
   },
   "devDependencies": {
+    "coveralls": "^3.0.2",
     "findit": "^1.2.0",
     "mkdirp": "^0.5.1",
-    "mocha": "*",
+    "mocha": "^5.2.0",
+    "nyc": "^12.0.2",
     "underscore": "^1.6.0"
   }
 }
